// nuxt-app/api/apiRoutes.ts
// Routes that provide data from api
export enum apiRoutes {
  CarouselImages = "/carousel-images",
  News = "/news",
  Programs = "/programs",
  ProgramVideos = "/program-videos",
  SchoolPrograms = "/school-programs",
  Pages = "/custom-pages",
  Videos = "/home-videos",
  Team = "/team",
  Documents = "/documents",
  Sponsors = "/sponsors",
  Editions = "/editions",
  SponsorsCategories = "/sponsor-categories",
  Residencies = "/residencies",
  SocialNetworks = "/social-networks",
  // wordpres routes
  WpCategories = "/categories",
}

export default apiRoutes